+++ src/Pim/Bundle/CatalogBundle/Doctrine/Common/Remover/GroupRemover.php
@@ -2,6 +2,7 @@
 
 namespace Pim\Bundle\CatalogBundle\Doctrine\Common\Remover;
 
+use Akeneo\Component\StorageUtils\Event\RemoveEvent;
 use Akeneo\Component\StorageUtils\Remover\RemoverInterface;
 use Akeneo\Component\StorageUtils\Remover\RemovingOptionsResolverInterface;
 use Doctrine\Common\Persistence\ObjectManager;
@@ -9,7 +10,6 @@ use Doctrine\Common\Util\ClassUtils;
 use Pim\Bundle\CatalogBundle\Event\GroupEvents;
 use Pim\Bundle\CatalogBundle\Model\GroupInterface;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
-use Symfony\Component\EventDispatcher\GenericEvent;
 
 /**
  * Group remover
@@ -60,11 +60,14 @@ class GroupRemover implements RemoverInterface
         }
 
         $options = $this->optionsResolver->resolveRemoveOptions($options);
-        $this->eventDispatcher->dispatch(GroupEvents::PRE_REMOVE, new GenericEvent($group));
+        $groupId = $group->getId();
+        $this->eventDispatcher->dispatch(GroupEvents::PRE_REMOVE, new RemoveEvent($group, $groupId));
 
         $this->objectManager->remove($group);
         if (true === $options['flush']) {
             $this->objectManager->flush();
         }
+
+        $this->eventDispatcher->dispatch(GroupEvents::POST_REMOVE, new RemoveEvent($group, $groupId));
     }
 }