+++ src/Pim/Bundle/EnrichBundle/Resources/public/js/product/form/sequential-edit.js
@@ -151,7 +151,6 @@ define(
                 var pending = objectSet[currentIndex + 2];
                 if (pending) {
                     setTimeout(function () {
-                        ProductManager.clear(pending);
                         ProductManager.get(pending);
                     }, 2000);
                 }